chore(poseidon-proof): move tests outside source folder
