+++ packages/poseidon-proof/src/config.ts
@@ -1,7 +1,4 @@
-import pkg from "../package.json"
-
-export const libraryName = pkg.name.split("/")[1]
-
+// eslint-disable-next-line import/prefer-default-export
 export const defaultSnarkArtifacts = {
     wasmFilePath: "https://zkkit.cedoor.dev/poseidon-proof.wasm",
     zkeyFilePath: "https://zkkit.cedoor.dev/poseidon-proof.zkey"
