+++ tsconfig.json
@@ -5,7 +5,6 @@
         "target": "ES5",
         "module": "ES2020",
         "moduleResolution": "node",
-        "resolveJsonModule": true,
         "esModuleInterop": true,
         "preserveConstEnums": true,
         "skipLibCheck": true,
